Update and rename Minimum wire cost required by a telecommunication company to Minimum wire cost required by a  company
The problem is finding the minimum wire cost required by a telecommunication company to connect 6 streets. The linesmen have given you the possible wiring they can perform and the distance between each pole when the wiring is done. As an engineer, you must give them the best possible wiring connections between poles so that the wiring cost is minimal to your company.